# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Controls one EC2 instance configured through environment variables.",
        "title": "EC2 Instance Controller API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/instance": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "Get the configured EC2 instance state and public IP addresses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.InstanceResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instance/state": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "Start or stop the configured EC2 instance",
                "parameters": [
                    {
                        "description": "Requested state transition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChangeStateRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.ChangeStateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ChangeStateRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "type": "string",
                    "enum": [
                        "start",
                        "stop"
                    ],
                    "example": "start"
                }
            }
        },
        "api.ChangeStateResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "start"
                },
                "current_state": {
                    "type": "string",
                    "example": "stopped"
                },
                "instance_id": {
                    "type": "string",
                    "example": "i-0123456789abcdef0"
                },
                "target_state": {
                    "type": "string",
                    "example": "pending"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
        "api.InstanceResponse": {
            "type": "object",
            "properties": {
                "instance_id": {
                    "type": "string",
                    "example": "i-0123456789abcdef0"
                },
                "public_ipv4": {
                    "type": "string",
                    "example": "203.0.113.10"
                },
                "public_ipv6": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "state": {
                    "type": "string",
                    "example": "running"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}